requirements.txt
gspread>=6.0.0
google-api-python-client>=2.131.0
google-auth>=2.33.0
requests>=2.31.0
